time series
- date_columns is the column for the claim admission date
- input_columns are the selected beneficiary ID rows from the dataframe
- output_variable is either the claim admission date or the day count 

I can't figure out a way to differentiate the beneficiary ID output variables as to display them differently by color. 

The only way that I can think to do this would be to use a loop since you have to adjust the plotting contingent on each BEN_ID. However, I know you're not a fan of loops so I avoided that. 

At the very least they'll be able to plot the claim day counts and claim payment amounts for each beneficiary ID or a combination of multiple IDs. It would be helpful to have color differentiation for up to 6-8 IDs before it becomes rather indistinguishable. I don't see it as a large issue as long as you can plot multiple IDs. Once we figure out how to select multiple IDs through the dropdown, I may be able to come up with some color differentiation device through matplotlib or another resource.